power: qpnp-smbcharger: add support for vfloat adjustments
The PMI8994 charger hardware has an analog control loop for regulating
float voltage during constant voltage charging. However, this control
loop has a 1% accuracy which is not acceptable for the fuel gauge
operation.

Add support for a software algorithm to automatically retrim the charger
float voltage regulation based on readings from the fuel gauge.

Note that the algorithm relies on periodic changes to the battery power
supply. The delta soc interrupt in Fuel Gauge causes a status change to
battery power supply and it's period is deemed good enough for invoking the
algorithm.

Change-Id: I85e9b1b6b337b29223b210e7f33623a287940df3
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>