power: qpnp-smbcharger: add support to enable AICL rerun
Currently we enable AICL rerun only when we want to limit
battery over-voltage oscillations seen while the charger is
headroom limited. However, there is a need to enable AICL
rerun all the time.

Add support for it using "qcom,force-aicl-rerun" property.

If the property is not set, AICL rerun and deglitch time will
be changed dynamically based on the battery voltage.

CRs-Fixed: 828718
Change-Id: Ic3e3f15793e76be5088f179751aa3cb0ca9bd4dd
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>